+++ docs/table-render/demo/advanced.tsx
@@ -1,6 +1,7 @@
 /**
  * transform: true
- * defaultShowCode: true
+ * defaultShowCode: false
+ * background: 'rgb(245,245,245)'
  */
 
 import { InfoCircleOutlined, PlusOutlined } from '@ant-design/icons';
@@ -32,7 +33,7 @@ const schema = {
       width: '25%',
     },
   },
-  'ui:labelWidth': 80,
+  labelWidth: 80,
 };
 
 const Demo = () => {
@@ -168,7 +169,7 @@ const Demo = () => {
   };
 
   return (
-    <div style={{ background: 'rgb(245,245,245)' }}>
+    <div>
       <Search
         schema={schema}
         displayType="row"