+++ packages/form-render/src/index.d.ts
@@ -1,6 +1,7 @@
 import { RuleItem } from 'async-validator';
 import * as React from 'react';
 
+export type { RuleItem } from 'async-validator';
 export type SchemaType =
   | 'string'
   | 'object'
@@ -12,7 +13,7 @@ export type SchemaType =
   | 'datetime'
   | (string & {});
 
-interface SchemaBase {
+export interface SchemaBase {
   type: SchemaType;
   title: string;
   description: string;
@@ -33,8 +34,10 @@ interface SchemaBase {
   placeholder: string;
   bind: false | string | string[];
   dependencies: string[];
-  min: number;
-  max: number;
+  /** 最小值，支持表达式 */
+  min: number | string;
+  /** 最大值，支持表达式 */
+  max: number | string;
   /** 是否禁用，支持 `'{{ formData.xxx === "" }}'` 形式的表达式 */
   disabled: boolean | string;
   /** 是否只读，支持 `'{{ formData.xxx === "" }}'` 形式的表达式 */
@@ -51,8 +54,10 @@ interface SchemaBase {
   extra: string;
   properties: Record<string, Schema>;
   items: Schema;
-  enum: Array<string | number>;
-  enumNames: Array<string | number>;
+  /** 多选，支持表达式 */
+  enum: Array<string | number> | string;
+  /** 多选label，支持表达式 */
+  enumNames: Array<string | number> | string;
   rules: RuleItem | RuleItem[];
   props: Record<string, any>;
 }
@@ -156,6 +161,10 @@ export interface FRProps {
   widgets?: any;
   /** 标签元素和输入元素的排列方式，column-分两行展示，row-同行展示，inline-自然顺排，默认`'column'` */
   displayType?: 'column' | 'row' | 'inline';
+  /** 表示是否显示 label 后面的冒号 */
+  colon?: boolean;
+  /** label 标签的文本对齐方式	 */
+  labelAlign?: 'right' | 'left';
   /** 只读模式 */
   readOnly?: boolean;
   /** 禁用模式 */